+++ tests/unit/Statement/StatementListTest.php
@@ -650,4 +650,14 @@ class StatementListTest extends \PHPUnit_Framework_TestCase {
 		);
 	}
 
+	public function testClear() {
+		$statement1 = $this->getStatement( 1, null );
+		$statement2 = $this->getStatement( 2, null );
+		$statements = new StatementList( $statement1, $statement2 );
+
+		$statements->clear();
+
+		$this->assertEquals( new StatementList(), $statements );
+	}
+
 }